Inline workspace state module
The indirection was net unhelpful.
